refactor: 타입 SMALLINT로 변경, order 예약어와 겹쳐서 column name orders로 변경

+++ wnp-server/src/main/java/com/wonnapark/wnpserver/domain/episode/EpisodeUrl.java
@@ -32,7 +32,7 @@ public class EpisodeUrl {
     @Column(name = "url", nullable = false)
     private String url;
 
-    @Column(name = "orders", nullable = false, columnDefinition = "TINYINT")
+    @Column(name = "orders", nullable = false, columnDefinition = "SMALLINT")
     private int order;
 
     @Column(name = "is_deleted")
